+++ evaluator/bool_func.go
@@ -1,7 +1,11 @@
 package evaluator
 
 import (
+	"errors"
+	"fmt"
+
 	"github.com/textwire/textwire/v2/ctx"
+	"github.com/textwire/textwire/v2/fail"
 	"github.com/textwire/textwire/v2/object"
 )
 
@@ -12,4 +16,22 @@ func boolBinaryFunc(_ *ctx.EvalCtx, receiver object.Object, _ ...object.Object) 
 	}
 
 	return &object.Int{Value: 0}, nil
+}
+
+// boolThenFunc returns the first argument if the receiver is true, the second argument or nil otherwise
+func boolThenFunc(_ *ctx.EvalCtx, receiver object.Object, args ...object.Object) (object.Object, error) {
+	if len(args) == 0 {
+		msg := fmt.Sprintf(fail.ErrFuncRequiresOneArg, "then", object.BOOL_OBJ)
+		return nil, errors.New(msg)
+	}
+
+	if isTruthy(receiver) {
+		return args[0], nil
+	}
+
+	if len(args) == 1 {
+		return &object.Nil{}, nil
+	}
+
+	return args[1], nil
 }